requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-cors==0.7.0
aiokafka==0.12.0
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
croniter==2.0.7
elastic-transport==8.15.1
elasticsearch==8.16.0
faust-streaming==0.11.3
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
intervaltree==3.1.0
kafka-python-ng==2.2.3
mode-streaming==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.27.0
opentracing==2.4.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
psycopg2-binary==2.9.10
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
terminaltables==3.1.10
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
venusian==3.1.0
yarl==1.18.3
